// client/src/types/Wall.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DoorState } from "./DoorState";
import type { WallType } from "./WallType";

export type Wall = { id: string, map_id: string, x1: number, y1: number, x2: number, y2: number, wall_type: WallType, door_state: DoorState, created_at: string, };